Fix: double into BigDecimal for monetary values. Start of MultiPartFile support.
Double into BigDecimal, fixing mistakes from startup of the project.
Start of MultiPartFile support for .pdf and .xml files.

+++ src/main/java/com/edu/ManagementPlayground/Dto/NotaFiscalResponseDto.java
@@ -2,9 +2,10 @@ package com.edu.ManagementPlayground.Dto;
 
 import com.edu.ManagementPlayground.Entity.NotaFiscal;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record NotaFiscalResponseDto(String numberIdentifier, LocalDate issueDate, double totalValue, String fileReference, long supplierIdentifier) {
+public record NotaFiscalResponseDto(String numberIdentifier, LocalDate issueDate, BigDecimal totalValue, String fileReference, long supplierIdentifier) {
 
     public static NotaFiscalResponseDto fromEntity(NotaFiscal notaFiscal){
         return new NotaFiscalResponseDto(
